+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## (unreleased)
+
+Features:
+  - reconnect(): Make logging to console optional
+  - close(): Add reconnectDelaySeconds property to close event
+
 ## 2.0.0 (2016-5-21)
 
 Features: